fix: исправить ошибки компиляции KMP после merge с origin/master
domain/commonMain:
- SettingsUseCase, SalarySettingUseCase, RouteUseCase:
  Dispatchers.IO → Dispatchers.Default (IO недоступен в KMP commonMain)
- DoubleUtil.kt: String.format()/«%.2f».format() (Java-only) заменены
  на ручную KMP-safe реализацию formatFixed2() и expandScientificNotation()

data_local:
- SearchRouteUseCase.kt: возвращён в androidMain (зависит от EntityString
  из core_android, которая недоступна в commonMain)
- SqlDelightRouteRepository.kt: добавлен .toMutableList() для locomotives/
  trains/passengers/photos (Route ожидает MutableList, map {} возвращает List)
- Train.sq: remoteObjectId TEXT NOT NULL → TEXT DEFAULT NULL
  (доменный объект Train имеет String?, SQL ожидал String)

data_remote/SyncManager.kt:
- selectMonthOfYear = currentMonthOfYear ?: listMonthOfYear.firstOrNull() ?: MonthOfYear()
  (UserSettings.selectMonthOfYear не допускает null)

iosApp/build.gradle.kts:
- Добавлена зависимость compose.materialIconsExtended
  (SettingsScreen/FormScreen/SalaryCalculationScreen используют Icons.AutoMirrored)
