+++ data_local/src/commonMain/kotlin/com/z_company/data_local/route/SearchRouteUseCase.kt
@@ -1,237 +1,2 @@
 package com.z_company.data_local.route
-
-import com.z_company.core.ResultState
-import com.z_company.core.util.EntityString
-import com.z_company.domain.entities.FilterSearch
-import com.z_company.domain.entities.RouteWithTag
-import com.z_company.domain.entities.SearchStateScreen
-import com.z_company.domain.entities.SearchTag
-import com.z_company.domain.entities.route.Route
-import com.z_company.domain.entities.route.UtilsForEntities.inTimePeriod
-import com.z_company.domain.repositories.RouteRepository
-import com.z_company.domain.util.addAllOrSkip
-import com.z_company.domain.util.addOrReplace
-import com.z_company.domain.util.addOrSkip
-import com.z_company.domain.util.splitBySpaceAndComma
-import kotlinx.coroutines.channels.awaitClose
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.channelFlow
-
-// Перемещён в commonMain: нет Android-зависимостей — только domain и core.
-class SearchRouteUseCase(val repository: RouteRepository) {
-    fun searchRoute(
-        value: String,
-        filter: FilterSearch,
-        preliminarySearch: Boolean,
-        entityString: EntityString
-    ): Flow<SearchStateScreen<List<RouteWithTag>>> =
-        channelFlow {
-            repository.loadRoutesAsStateFlow().collect { result ->
-                when (result) {
-                    is ResultState.Loading -> trySend(SearchStateScreen.Loading())
-
-                    is ResultState.Success -> result.data.let { resultList ->
-                        val sortedList = mutableListOf<RouteWithTag>()
-
-                        resultList.forEach { route ->
-                            if (route.inTimePeriod(filter.timePeriod)) {
-                                if (filter.generalData.second) {
-                                    sortedList.addAllOrSkip(searchInRouteData(route, value, entityString))
-                                }
-                                if (filter.locoData.second) {
-                                    sortedList.addAllOrSkip(searchInLocoList(route, value, entityString))
-                                }
-                                if (filter.trainData.second) {
-                                    sortedList.addAllOrSkip(searchInTrainList(route, value, entityString))
-                                }
-                                if (filter.passengerData.second) {
-                                    sortedList.addAllOrSkip(searchInPassengerList(route, value, entityString))
-                                }
-                                if (filter.notesData.second) {
-                                    sortedList.addAllOrSkip(searchInNotes(route, value))
-                                }
-                            }
-                        }
-                        if (preliminarySearch) {
-                            val hintsList = mutableListOf<String>()
-
-                            sortedList.forEach { routeWithTag: RouteWithTag ->
-                                val text = when (routeWithTag.tag) {
-                                    SearchTag.BASIC_DATA -> {
-                                        StringBuilder(entityString.basicDataStr(routeWithTag.route.basicData))
-                                    }
-
-                                    SearchTag.LOCO -> {
-                                        val text = StringBuilder()
-                                        routeWithTag.route.locomotives.forEach { loco ->
-                                            text.append("${entityString.locomotiveStr(loco)} ")
-                                        }
-                                        text
-                                    }
-
-                                    SearchTag.TRAIN -> {
-                                        val text = StringBuilder()
-                                        routeWithTag.route.trains.forEach { train ->
-                                            text.append("${entityString.trainStr(train)} ")
-                                        }
-                                        text
-                                    }
-
-                                    SearchTag.PASSENGER -> {
-                                        val text = StringBuilder()
-                                        routeWithTag.route.passengers.forEach { passenger ->
-                                            text.append(entityString.passengerStr(passenger))
-                                        }
-                                        text
-                                    }
-
-                                    SearchTag.NOTES -> {
-                                        StringBuilder(routeWithTag.route.basicData.notes.toString())
-                                    }
-                                }
-
-                                val listStrings = text.toString().splitBySpaceAndComma()
-                                val valueString = value.splitBySpaceAndComma().last()
-
-                                listStrings.forEachIndexed { index, s ->
-                                    if (s == valueString && index != listStrings.lastIndex) {
-                                        hintsList.addOrSkip(listStrings[index + 1])
-                                    } else if (s.contains(value)) {
-                                        hintsList.addOrSkip(s)
-                                    }
-                                }
-                            }
-                            trySend(
-                                SearchStateScreen.Input(hintsList)
-                            )
-                        } else {
-                            trySend(SearchStateScreen.Success(sortedList.sortedBy { it.route.basicData.timeStartWork }))
-                        }
-                    }
-
-                    is ResultState.Error -> trySend(SearchStateScreen.Failure(result.entity))
-                }
-            }
-            awaitClose()
-        }
-
-
-    private fun searchInRouteData(
-        route: Route,
-        value: String,
-        entityString: EntityString
-    ): MutableList<RouteWithTag> {
-        val filteredList = mutableListOf<RouteWithTag>()
-
-        val valueList = value.splitBySpaceAndComma()
-        var respond = true
-
-        valueList.forEach { s ->
-            if (!entityString.basicDataStr(route.basicData).contains(s, ignoreCase = true)) {
-                respond = false
-            }
-        }
-
-        if (respond) {
-            filteredList.addOrReplace(RouteWithTag(SearchTag.BASIC_DATA, route))
-        }
-        return filteredList
-    }
-
-    private fun searchInLocoList(
-        route: Route,
-        value: String,
-        entityString: EntityString
-    ): MutableList<RouteWithTag> {
-        val filteredList = mutableListOf<RouteWithTag>()
-        route.locomotives.forEach { loco ->
-
-            val valueList = value.splitBySpaceAndComma()
-            var respond = true
-
-            valueList.forEach { s ->
-                if (!entityString.locomotiveStr(loco).contains(s, ignoreCase = true)) {
-                    respond = false
-                }
-            }
-
-            if (respond) {
-                filteredList.addOrReplace(RouteWithTag(SearchTag.LOCO, route))
-            }
-        }
-        return filteredList
-    }
-
-    private fun searchInTrainList(
-        route: Route,
-        value: String,
-        entityString: EntityString
-    ): MutableList<RouteWithTag> {
-        val filteredList = mutableListOf<RouteWithTag>()
-
-        route.trains.forEach { train ->
-            val valueList = value.splitBySpaceAndComma()
-            var respond = true
-
-            valueList.forEach { s ->
-                if (!entityString.trainStr(train).contains(s, ignoreCase = true)) {
-                    respond = false
-                }
-            }
-
-            if (respond) {
-                filteredList.addOrReplace(RouteWithTag(SearchTag.TRAIN, route))
-            }
-        }
-        return filteredList
-    }
-
-    private fun searchInPassengerList(
-        route: Route,
-        value: String,
-        entityString: EntityString
-    ): MutableList<RouteWithTag> {
-        val filteredList = mutableListOf<RouteWithTag>()
-
-        route.passengers.forEach { passenger ->
-
-            val valueList = value.splitBySpaceAndComma()
-            var respond = true
-
-            valueList.forEach { s ->
-                if (!entityString.passengerStr(passenger).contains(s, ignoreCase = true)) {
-                    respond = false
-                }
-            }
-
-            if (respond) {
-                filteredList.addOrReplace(RouteWithTag(SearchTag.PASSENGER, route))
-            }
-        }
-        return filteredList
-    }
-
-    private fun searchInNotes(
-        route: Route,
-        value: String
-    ): MutableList<RouteWithTag> {
-        val filteredList = mutableListOf<RouteWithTag>()
-
-        route.basicData.notes?.let { note ->
-
-            val valueList = value.splitBySpaceAndComma()
-            var respond = true
-
-            valueList.forEach { s ->
-                if (!note.contains(s, ignoreCase = true)) {
-                    respond = false
-                }
-            }
-
-            if (respond) {
-                filteredList.addOrSkip(RouteWithTag(SearchTag.NOTES, route))
-            }
-        }
-        return filteredList
-    }
-}
+// Перемещён в androidMain: зависит от EntityString из core_android (Android-only)
